finished cleaning up scale drawing and calculations

+++ app/src/main/java/camp/scottlikesto/www/magnificationoverlay/Scale.java
@@ -2,10 +2,10 @@ package camp.scottlikesto.www.magnificationoverlay;
 
 public class Scale {
     public int magnification;
-    public int length;
+    public double length;
     public String unit;
 
-    public Scale(int _mag, int _length, String _unit) {
+    public Scale(int _mag, double _length, String _unit) {
         magnification = _mag;
         length = _length;
         unit = _unit;
